+++ src/Service/Kinesis/src/Input/AddTagsToStreamInput.php
@@ -15,8 +15,6 @@ final class AddTagsToStreamInput extends Input
     /**
      * The name of the stream.
      *
-     * @required
-     *
      * @var string|null
      */
     private $streamName;
@@ -31,9 +29,17 @@ final class AddTagsToStreamInput extends Input
     private $tags;
 
     /**
+     * The ARN of the stream.
+     *
+     * @var string|null
+     */
+    private $streamArn;
+
+    /**
      * @param array{
      *   StreamName?: string,
      *   Tags?: array<string, string>,
+     *   StreamARN?: string,
      *   @region?: string,
      * } $input
      */
@@ -41,12 +47,18 @@ final class AddTagsToStreamInput extends Input
     {
         $this->streamName = $input['StreamName'] ?? null;
         $this->tags = $input['Tags'] ?? null;
+        $this->streamArn = $input['StreamARN'] ?? null;
         parent::__construct($input);
     }
 
     public static function create($input): self
     {
         return $input instanceof self ? $input : new self($input);
+    }
+
+    public function getStreamArn(): ?string
+    {
+        return $this->streamArn;
     }
 
     public function getStreamName(): ?string
@@ -87,6 +99,13 @@ final class AddTagsToStreamInput extends Input
         return new Request('POST', $uriString, $query, $headers, StreamFactory::create($body));
     }
 
+    public function setStreamArn(?string $value): self
+    {
+        $this->streamArn = $value;
+
+        return $this;
+    }
+
     public function setStreamName(?string $value): self
     {
         $this->streamName = $value;
@@ -107,10 +126,9 @@ final class AddTagsToStreamInput extends Input
     private function requestBody(): array
     {
         $payload = [];
-        if (null === $v = $this->streamName) {
-            throw new InvalidArgument(sprintf('Missing parameter "StreamName" for "%s". The value cannot be null.', __CLASS__));
+        if (null !== $v = $this->streamName) {
+            $payload['StreamName'] = $v;
         }
-        $payload['StreamName'] = $v;
         if (null === $v = $this->tags) {
             throw new InvalidArgument(sprintf('Missing parameter "Tags" for "%s". The value cannot be null.', __CLASS__));
         }
@@ -122,6 +140,9 @@ final class AddTagsToStreamInput extends Input
             foreach ($v as $name => $mv) {
                 $payload['Tags'][$name] = $mv;
             }
+        }
+        if (null !== $v = $this->streamArn) {
+            $payload['StreamARN'] = $v;
         }
 
         return $payload;