+++ src/Service/Kinesis/src/Input/GetShardIteratorInput.php
@@ -16,8 +16,6 @@ final class GetShardIteratorInput extends Input
     /**
      * The name of the Amazon Kinesis data stream.
      *
-     * @required
-     *
      * @var string|null
      */
     private $streamName;
@@ -60,12 +58,20 @@ final class GetShardIteratorInput extends Input
     private $timestamp;
 
     /**
+     * The ARN of the stream.
+     *
+     * @var string|null
+     */
+    private $streamArn;
+
+    /**
      * @param array{
      *   StreamName?: string,
      *   ShardId?: string,
      *   ShardIteratorType?: ShardIteratorType::*,
      *   StartingSequenceNumber?: string,
      *   Timestamp?: \DateTimeImmutable|string,
+     *   StreamARN?: string,
      *   @region?: string,
      * } $input
      */
@@ -76,6 +82,7 @@ final class GetShardIteratorInput extends Input
         $this->shardIteratorType = $input['ShardIteratorType'] ?? null;
         $this->startingSequenceNumber = $input['StartingSequenceNumber'] ?? null;
         $this->timestamp = !isset($input['Timestamp']) ? null : ($input['Timestamp'] instanceof \DateTimeImmutable ? $input['Timestamp'] : new \DateTimeImmutable($input['Timestamp']));
+        $this->streamArn = $input['StreamARN'] ?? null;
         parent::__construct($input);
     }
 
@@ -100,6 +107,11 @@ final class GetShardIteratorInput extends Input
     public function getStartingSequenceNumber(): ?string
     {
         return $this->startingSequenceNumber;
+    }
+
+    public function getStreamArn(): ?string
+    {
+        return $this->streamArn;
     }
 
     public function getStreamName(): ?string
@@ -161,6 +173,13 @@ final class GetShardIteratorInput extends Input
         return $this;
     }
 
+    public function setStreamArn(?string $value): self
+    {
+        $this->streamArn = $value;
+
+        return $this;
+    }
+
     public function setStreamName(?string $value): self
     {
         $this->streamName = $value;
@@ -178,10 +197,9 @@ final class GetShardIteratorInput extends Input
     private function requestBody(): array
     {
         $payload = [];
-        if (null === $v = $this->streamName) {
-            throw new InvalidArgument(sprintf('Missing parameter "StreamName" for "%s". The value cannot be null.', __CLASS__));
+        if (null !== $v = $this->streamName) {
+            $payload['StreamName'] = $v;
         }
-        $payload['StreamName'] = $v;
         if (null === $v = $this->shardId) {
             throw new InvalidArgument(sprintf('Missing parameter "ShardId" for "%s". The value cannot be null.', __CLASS__));
         }
@@ -198,6 +216,9 @@ final class GetShardIteratorInput extends Input
         }
         if (null !== $v = $this->timestamp) {
             $payload['Timestamp'] = $v->format(\DateTimeInterface::ATOM);
+        }
+        if (null !== $v = $this->streamArn) {
+            $payload['StreamARN'] = $v;
         }
 
         return $payload;