+++ src/Service/Kinesis/src/Input/MergeShardsInput.php
@@ -15,8 +15,6 @@ final class MergeShardsInput extends Input
     /**
      * The name of the stream for the merge.
      *
-     * @required
-     *
      * @var string|null
      */
     private $streamName;
@@ -40,10 +38,18 @@ final class MergeShardsInput extends Input
     private $adjacentShardToMerge;
 
     /**
+     * The ARN of the stream.
+     *
+     * @var string|null
+     */
+    private $streamArn;
+
+    /**
      * @param array{
      *   StreamName?: string,
      *   ShardToMerge?: string,
      *   AdjacentShardToMerge?: string,
+     *   StreamARN?: string,
      *   @region?: string,
      * } $input
      */
@@ -52,6 +58,7 @@ final class MergeShardsInput extends Input
         $this->streamName = $input['StreamName'] ?? null;
         $this->shardToMerge = $input['ShardToMerge'] ?? null;
         $this->adjacentShardToMerge = $input['AdjacentShardToMerge'] ?? null;
+        $this->streamArn = $input['StreamARN'] ?? null;
         parent::__construct($input);
     }
 
@@ -68,6 +75,11 @@ final class MergeShardsInput extends Input
     public function getShardToMerge(): ?string
     {
         return $this->shardToMerge;
+    }
+
+    public function getStreamArn(): ?string
+    {
+        return $this->streamArn;
     }
 
     public function getStreamName(): ?string
@@ -114,6 +126,13 @@ final class MergeShardsInput extends Input
         return $this;
     }
 
+    public function setStreamArn(?string $value): self
+    {
+        $this->streamArn = $value;
+
+        return $this;
+    }
+
     public function setStreamName(?string $value): self
     {
         $this->streamName = $value;
@@ -124,10 +143,9 @@ final class MergeShardsInput extends Input
     private function requestBody(): array
     {
         $payload = [];
-        if (null === $v = $this->streamName) {
-            throw new InvalidArgument(sprintf('Missing parameter "StreamName" for "%s". The value cannot be null.', __CLASS__));
+        if (null !== $v = $this->streamName) {
+            $payload['StreamName'] = $v;
         }
-        $payload['StreamName'] = $v;
         if (null === $v = $this->shardToMerge) {
             throw new InvalidArgument(sprintf('Missing parameter "ShardToMerge" for "%s". The value cannot be null.', __CLASS__));
         }
@@ -136,6 +154,9 @@ final class MergeShardsInput extends Input
             throw new InvalidArgument(sprintf('Missing parameter "AdjacentShardToMerge" for "%s". The value cannot be null.', __CLASS__));
         }
         $payload['AdjacentShardToMerge'] = $v;
+        if (null !== $v = $this->streamArn) {
+            $payload['StreamARN'] = $v;
+        }
 
         return $payload;
     }