+++ src/Service/Kinesis/src/Result/ListStreamsOutput.php
@@ -7,11 +7,13 @@ use AsyncAws\Core\Response;
 use AsyncAws\Core\Result;
 use AsyncAws\Kinesis\Input\ListStreamsInput;
 use AsyncAws\Kinesis\KinesisClient;
+use AsyncAws\Kinesis\ValueObject\StreamModeDetails;
+use AsyncAws\Kinesis\ValueObject\StreamSummary;
 
 /**
  * Represents the output for `ListStreams`.
  *
- * @implements \IteratorAggregate<string>
+ * @implements \IteratorAggregate<string|StreamSummary>
  */
 class ListStreamsOutput extends Result implements \IteratorAggregate
 {
@@ -25,6 +27,10 @@ class ListStreamsOutput extends Result implements \IteratorAggregate
      */
     private $hasMoreStreams;
 
+    private $nextToken;
+
+    private $streamSummaries;
+
     public function getHasMoreStreams(): bool
     {
         $this->initialize();
@@ -33,13 +39,48 @@ class ListStreamsOutput extends Result implements \IteratorAggregate
     }
 
     /**
-     * Iterates over StreamNames.
+     * Iterates over StreamNames and StreamSummaries.
      *
-     * @return \Traversable<string>
+     * @return \Traversable<string|StreamSummary>
      */
     public function getIterator(): \Traversable
     {
-        yield from $this->getStreamNames();
+        $client = $this->awsClient;
+        if (!$client instanceof KinesisClient) {
+            throw new InvalidArgument('missing client injected in paginated result');
+        }
+        if (!$this->input instanceof ListStreamsInput) {
+            throw new InvalidArgument('missing last request injected in paginated result');
+        }
+        $input = clone $this->input;
+        $page = $this;
+        while (true) {
+            $page->initialize();
+            if ($page->hasMoreStreams) {
+                $input->setNextToken($page->nextToken);
+
+                $this->registerPrefetch($nextPage = $client->listStreams($input));
+            } else {
+                $nextPage = null;
+            }
+
+            yield from $page->getStreamNames(true);
+            yield from $page->getStreamSummaries(true);
+
+            if (null === $nextPage) {
+                break;
+            }
+
+            $this->unregisterPrefetch($nextPage);
+            $page = $nextPage;
+        }
+    }
+
+    public function getNextToken(): ?string
+    {
+        $this->initialize();
+
+        return $this->nextToken;
     }
 
     /**
@@ -68,7 +109,7 @@ class ListStreamsOutput extends Result implements \IteratorAggregate
         while (true) {
             $page->initialize();
             if ($page->hasMoreStreams) {
-                $input->setExclusiveStartStreamName(\array_slice($page->streamNames, -1)[0]);
+                $input->setNextToken($page->nextToken);
 
                 $this->registerPrefetch($nextPage = $client->listStreams($input));
             } else {
@@ -86,12 +127,65 @@ class ListStreamsOutput extends Result implements \IteratorAggregate
         }
     }
 
+    /**
+     * @param bool $currentPageOnly When true, iterates over items of the current page. Otherwise also fetch items in the next pages.
+     *
+     * @return iterable<StreamSummary>
+     */
+    public function getStreamSummaries(bool $currentPageOnly = false): iterable
+    {
+        if ($currentPageOnly) {
+            $this->initialize();
+            yield from $this->streamSummaries;
+
+            return;
+        }
+
+        $client = $this->awsClient;
+        if (!$client instanceof KinesisClient) {
+            throw new InvalidArgument('missing client injected in paginated result');
+        }
+        if (!$this->input instanceof ListStreamsInput) {
+            throw new InvalidArgument('missing last request injected in paginated result');
+        }
+        $input = clone $this->input;
+        $page = $this;
+        while (true) {
+            $page->initialize();
+            if ($page->hasMoreStreams) {
+                $input->setNextToken($page->nextToken);
+
+                $this->registerPrefetch($nextPage = $client->listStreams($input));
+            } else {
+                $nextPage = null;
+            }
+
+            yield from $page->streamSummaries;
+
+            if (null === $nextPage) {
+                break;
+            }
+
+            $this->unregisterPrefetch($nextPage);
+            $page = $nextPage;
+        }
+    }
+
     protected function populateResult(Response $response): void
     {
         $data = $response->toArray();
 
         $this->streamNames = $this->populateResultStreamNameList($data['StreamNames']);
         $this->hasMoreStreams = filter_var($data['HasMoreStreams'], \FILTER_VALIDATE_BOOLEAN);
+        $this->nextToken = isset($data['NextToken']) ? (string) $data['NextToken'] : null;
+        $this->streamSummaries = empty($data['StreamSummaries']) ? [] : $this->populateResultStreamSummaryList($data['StreamSummaries']);
+    }
+
+    private function populateResultStreamModeDetails(array $json): StreamModeDetails
+    {
+        return new StreamModeDetails([
+            'StreamMode' => (string) $json['StreamMode'],
+        ]);
     }
 
     /**
@@ -105,6 +199,30 @@ class ListStreamsOutput extends Result implements \IteratorAggregate
             if (null !== $a) {
                 $items[] = $a;
             }
+        }
+
+        return $items;
+    }
+
+    private function populateResultStreamSummary(array $json): StreamSummary
+    {
+        return new StreamSummary([
+            'StreamName' => (string) $json['StreamName'],
+            'StreamARN' => (string) $json['StreamARN'],
+            'StreamStatus' => (string) $json['StreamStatus'],
+            'StreamModeDetails' => empty($json['StreamModeDetails']) ? null : $this->populateResultStreamModeDetails($json['StreamModeDetails']),
+            'StreamCreationTimestamp' => (isset($json['StreamCreationTimestamp']) && ($d = \DateTimeImmutable::createFromFormat('U.u', sprintf('%.6F', $json['StreamCreationTimestamp'])))) ? $d : null,
+        ]);
+    }
+
+    /**
+     * @return StreamSummary[]
+     */
+    private function populateResultStreamSummaryList(array $json): array
+    {
+        $items = [];
+        foreach ($json as $item) {
+            $items[] = $this->populateResultStreamSummary($item);
         }
 
         return $items;