+++ src/Service/Translate/src/Input/TranslateTextRequest.php
@@ -12,8 +12,8 @@ use AsyncAws\Translate\ValueObject\TranslationSettings;
 final class TranslateTextRequest extends Input
 {
     /**
-     * The text to translate. The text string can be a maximum of 5,000 bytes long. Depending on your character set, this
-     * may be fewer than 5,000 characters.
+     * The text to translate. The text string can be a maximum of 10,000 bytes long. Depending on your character set, this
+     * may be fewer than 10,000 characters.
      *
      * @required
      *